LayoutTests/imported/w3c/web-platform-tests/paint-timing/fcp-only/fcp-invisible-3d-rotate-descendant-expected.txt

Text

FAIL First contentful paint fires due to its ancestor getting rotating into view. assert_equals: First contentful paint marked too early.  expected 0 but got 1
